Added: #1 Modular thread crawling per year
Now it is possible to crawl threads for a dedicated year by simply executing a .py - file.
These crawler files also have an update mechanism integrated.
Whenever it iterates over a, already existing, collection, it will check its values and, depending on it, by respecting an integrated tolerance, it will drop that collection and write it a new, or not :)