Added: #1 Modular comment crawler per year
As in the previous comment it is now possible to crawl every single comment, in full-text, on a year - basis.
An update mechanism is to be implemented as next step, because at the moment the comment creation will be skipped, whenever there already exists a collection with the same name.