Added: #1 Improved comments crawler
Now, whenever "crawl_Threads_All_20XX" will be executed it additionally checks if the amount of comments has changed.
Whenever that happens the appropriate collection in the comments database gets dropped and is to be recrawled anew by using either "diff-scripts" or "crawl_Comments_All_20XX" scripts.
This is the only way we can guarantee that nearly all comments are up to date in the database