Changed: #19, #21 Question tier distribution calculation error
I have fixed a bad bug, which caused to generate data which is unlogical..
The chart contained other data than the csv files (in terms of amount, etc..).
This is fixed now!
It is not a very nice code, because we loop twice over the same data, but it works rock solid.