Changed: #45, #36 Big rework + creation of flask web server
Now our extension is able to logon via oauth2 with rest calls - all in python.
After successful login the index.html gets reserved.
Now we have a all-in-one-solution which is platform independend.
What is now left to be done is to load data from the user, display appropriate threads and making posting behaviour work.